requirements.txt
allure-pytest
faker
requests
pytest
voluptuous
pytest-voluptuous